+++ src/main/java/net/javacrumbs/restfire/httpcomponents/HttpComponentsRequestFactory.java
@@ -18,7 +18,6 @@ package net.javacrumbs.restfire.httpcomponents;
 import net.javacrumbs.restfire.RequestBuilder;
 import net.javacrumbs.restfire.RequestFactory;
 import net.javacrumbs.restfire.RequestProcessor;
-import net.javacrumbs.restfire.ResponseValidator;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpGet;